# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "authorization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "authorization",
                "parameters": [
                    {
                        "description": "User email and password",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.tAuthorization"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "description": "logout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/otp": {
            "post": {
                "description": "send code to email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "send to email one time password",
                "parameters": [
                    {
                        "description": "User email",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.tRequestOTP"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/tournaments": {
            "get": {
                "description": "все турниры",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "guest"
                ],
                "summary": "все турниры",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.tGetTorunamentsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "description": "user info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "user info",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/players": {
            "get": {
                "description": "Все игроки",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user players"
                ],
                "summary": "Все игроки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.tGetPlayersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Добавить игрока",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user players"
                ],
                "summary": "Добавить игрока",
                "parameters": [
                    {
                        "description": "player",
                        "name": "player",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.tNewPlayer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/rest.tPlayer"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/players/{player_id}": {
            "put": {
                "description": "обновить игрока",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user players"
                ],
                "summary": "обновить игрока",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "player id",
                        "name": "player_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "player",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.tUpdatePlayer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.tPlayer"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/players/{player_id}/upload": {
            "put": {
                "description": "загрузка файлов игрока",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user players"
                ],
                "summary": "загрузка файлов игрока",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "player id",
                        "name": "player_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "фотография",
                        "name": "photo_file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/teams": {
            "get": {
                "description": "команды пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user team"
                ],
                "summary": "команды пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.tGetTeamsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "создать команду",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user team"
                ],
                "summary": "создать команду",
                "parameters": [
                    {
                        "description": "team",
                        "name": "tournamet",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.tCreateTeam"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/rest.tTeam"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/teams/{team_id}": {
            "get": {
                "description": "информация команды пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user team"
                ],
                "summary": "информация команды пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "team id",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.tGetTeamResponse"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "обновление команды пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user team"
                ],
                "summary": "обновление команды пользователя",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "team id",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "team",
                        "name": "team",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.tUpdTeamRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.tUpdTeamResponse"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/teams/{team_id}/applications": {
            "get": {
                "description": "заявки команды",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user team"
                ],
                "summary": "заявки команды",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "team id",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.tGetApplicationsTeamResponse"
                        }
                    },
                    "400": {
                        "description": "команда не найдена"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "подать заявку",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user team"
                ],
                "summary": "подать заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "team id",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "application",
                        "name": "application",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.tNewApplicationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "заявка создана",
                        "schema": {
                            "$ref": "#/definitions/rest.tNewApplicationResponse"
                        }
                    },
                    "400": {
                        "description": "не корректный запрос"
                    },
                    "409": {
                        "description": "заявка\tуже\tбыла создана ранее"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/teams/{team_id}/applications/{application_id}": {
            "get": {
                "description": "заявка команды",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user team"
                ],
                "summary": "заявка команды",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "team id",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "application id",
                        "name": "application_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.tGetApplicationResponse"
                        }
                    },
                    "400": {
                        "description": "не корректный запрос"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "изменить заявку",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user team"
                ],
                "summary": "изменить заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "team id",
                        "name": "team_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "application\tid",
                        "name": "application_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "application status",
                        "name": "application",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.tUpdApplicationStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.tUpdApplicationResponse"
                        }
                    },
                    "204": {
                        "description": "заявка не найдена"
                    },
                    "400": {
                        "description": "не найден или не может изменить"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/user/tournaments/{tournament_id}/applications": {
            "get": {
                "description": "заявки на турнир",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user tournament"
                ],
                "summary": "заявки на турнир",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "tournament id",
                        "name": "tournament_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.tGetTournamentApplicationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "rest.tAuthorization": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "otp": {
                    "description": "Password string `json:\"password\"`",
                    "type": "string"
                }
            }
        },
...
        "rest.tGetPlayersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.tPlayer"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/rest.pagination"
                }
            }
        },
...
        "rest.tGetTorunamentsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/rest.tTournament"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/rest.pagination"
                }
            }
        },
...
        "rest.tNewPlayer": {
            "type": "object",
            "properties": {
                "b_day": {
                    "type": "string",
                    "example": "2024-12-31"
                },
                "firstname": {
                    "type": "string"
                },
                "lastname": {
                    "type": "string"
                },
                "secondname": {
                    "type": "string"
                }
            }
        },
...
        "rest.tRequestOTP": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
    },
...
}